web/backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.32
supabase>=2.7
pyjwt>=2.9
openai>=1.50
pydantic>=2.9
pydantic-settings>=2.6
python-dotenv>=1.0
